better refactoring of components
