+++ src/app/page.tsx
@@ -1,9 +1,5 @@
-import {RepositoryForm} from "@/components/Forms/RepositoryForm";
-
 export default function Home() {
   return (
-    <div className="flex flex-col items-center justify-center h-screen">
-      <RepositoryForm />
-    </div>
+    <div>Welcome to devoter</div>
   );
 }
