+++ src/app/submit-repo/page.tsx
@@ -1,0 +1,9 @@
+import { SubmitRepoPageContent } from '@/components/pages/submit-repo';
+
+export default function SubmitRepoPage() {
+  return (
+    <div className="flex flex-col items-center justify-center h-screen">
+      <SubmitRepoPageContent />
+    </div>
+  );
+}
